Adding functionality so that when an automation trigger is deleted its test inputs will also be deleted.
